+++ packages/openapi-generator/src/options/flags.ts
@@ -2,7 +2,7 @@ import { resolve, extname } from 'path';
 import { existsSync, lstatSync } from 'fs';
 import { flags } from '@oclif/command';
 
-export const openapi_flags = {
+export const generatorFlags = {
   input: flags.string({
     char: 'i',
     description:
@@ -93,8 +93,25 @@ export const openapi_flags = {
   config: flags.string({
     char: 'c',
     parse: input => resolve(input),
-    description: 'Set the path to the config.json file for generation.',
+    description:
+      'Set the path to the a file containing the options for generation. If the config flag is used, no other flags can be used. If a directory is passed, a `config.json` file is read from this directory.',
     helpValue: '<path/to/config.json>',
-    required: false
+    required: false,
+    exclusive: [
+      'input',
+      'outputDir',
+      'transpile',
+      'include',
+      'overwrite',
+      'clearOutputDir',
+      'skipValidation',
+      'tsConfig',
+      'packageJson',
+      'verbose',
+      'optionsPerService',
+      'packageVersion',
+      'readme',
+      'metadata'
+    ]
   })
 };