+++ packages/openapi-generator/src/options/generator-options.ts
@@ -1,6 +1,6 @@
 import Parser from '@oclif/parser';
 import OpenApiGenerator from '../cli';
-import { openapi_flags } from './flags';
+import { generatorFlags } from './flags';
 
 /**
  * Options that can be used to configure the generation when using the generator programmatically.
@@ -44,13 +44,16 @@ export type ParsedGeneratorOptions = typeof OpenApiGenerator extends Parser.Inpu
 export function parseGeneratorOptions(
   options: GeneratorOptions
 ): ParsedGeneratorOptions {
-  return Object.entries(openapi_flags).reduce((parsedOptions, [name, flag]) => {
-    const value = options[name];
-    if (typeof value !== 'undefined') {
-      parsedOptions[name] = flag.parse(value as never, undefined);
-    } else if (typeof flag.default !== undefined) {
-      parsedOptions[name] = flag.default;
-    }
-    return parsedOptions;
-  }, {} as ParsedGeneratorOptions);
+  return Object.entries(generatorFlags).reduce(
+    (parsedOptions, [name, flag]) => {
+      const value = options[name];
+      if (typeof value !== 'undefined') {
+        parsedOptions[name] = flag.parse(value as never, undefined);
+      } else if (typeof flag.default !== undefined) {
+        parsedOptions[name] = flag.default;
+      }
+      return parsedOptions;
+    },
+    {} as ParsedGeneratorOptions
+  );
 }